Move features.rs to device module
